requirements.txt
# ============================================================
# DATAWAREHOUSE PROJECT - Requirements
# Capa: Bronce (Ingesta desde Google Drive + PostgreSQL)
# ============================================================

# Google Drive & Sheets
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.131.0
gspread==6.1.2

# Excel / Data
openpyxl==3.1.3
pandas==2.2.2
xlrd==2.0.1

# PostgreSQL
psycopg2-binary==2.9.9
sqlalchemy==2.0.30

# Utilidades
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.3.0
loguru==0.7.2
tenacity==8.3.0
tqdm==4.66.4

# Testing
pytest==8.2.2
pytest-cov==5.0.0

# Opcional: scheduling local
schedule==1.2.2
